Starts implementing the navicon
It’s not done yet, but all the view controllers the navicon will use
are here.

Adds a folder for supporting files. Right now that folder contains .h
and .m files for SWRevealViewController (a UIViewController subclass),
plus info.plist.

Adds potted_plant.jpg. This is perhaps the most crucial step in
developing an app about plants. Become one with the plant. Admire its
beauty. The fibers of this plant intertwine exquisitely with the fibers
of the cosmos themselves. Perhaps we all live in pots of one kind or
another; ours is just a swirling mass of blue and green, hurtling
through the vast unknown. You know this plant. It speaks to you, like a
mother speaks to her beloved child. This plant knows you. It has roots.
It is your roots. You are home.

Also adds LaunchScreen.storyboard featuring a picture of said potted
plant.