Adds the hamburger button!
And there was much rejoicing.

This commit adds the CocoaPods files required to use SlideMenu, which
is the library we are using to create the navicon. I also wired up the
navicon in Main.storyboard so that it slides out to display some empty
cells.

Also, perhaps most importantly of all, potted_plant.jpg returns to the
repo. We’ve missed you old friend.