Broken pods and broken dreams
THIS CODE DOES NOT BUILD

Turns out Google’s GTMOAuth2 pod is deprecated as of a few months ago
so we’re gonna have to use their new version called GTMAppAuth instead.
This code is a snapshot of the project before I start porting to
GTMAppAuth.

Thing is, that new version is only documented for Objective-C and we’re
using Swift files…